+++ src/_.js
@@ -8,7 +8,7 @@ var µ = µ || {};
     µ.global = global;
 
     µ.DEBUG = false;
-    µ.VERSION = '0.1.0';
+    µ.VERSION = '0.1.1';
 
     µ.apply = function (object, config, defaults) {
         if (defaults) {