+++ src/format.js
@@ -136,6 +136,23 @@
         },
 
         /**
+         * format number by adding thousands separaters and significant digits while rounding
+         */
+        num: function (number, decimals, decPoint, thousandsSep) {
+            decimals = isNaN(decimals) ? 2 : Math.abs(decimals);
+            decPoint = (decPoint === undefined) ? '.' : decPoint;
+            thousandsSep = (thousandsSep === undefined) ? ',' : thousandsSep;
+
+            var sign = number < 0 ? '-' : '';
+            number = Math.abs(+number || 0);
+
+            var intPart = parseInt(number.toFixed(decimals), 10) + '';
+            var j = intPart.length > 3 ? intPart.length % 3 : 0;
+
+            return sign + (j ? intPart.substr(0, j) + thousandsSep : '') + intPart.substr(j).replace(/(\d{3})(?=\d)/g, '$1' + thousandsSep) + (decimals ? decPoint + Math.abs(number - intPart).toFixed(decimals).slice(2) : '');
+        },
+
+        /**
          * Validate that a value is numeric and convert it to a number if necessary. Returns the specified default value if
          * it is not.